+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Created By: BENGUERGOURA Oussama & HADJERSI Mohamed | 20/10/2024
 
 ### Added
 
+* Creating `BarSide`component with multiple `BarSideMenu` and `BarSideItem`. [GTS-PRINT#32](https://quire.io/w/GTS-PRINT31/2)
 
 ### Changed